+++ src/test/java/usertest/CreateUserTest.java
@@ -8,46 +8,46 @@ import org.junit.Before;
 import org.junit.Test;
 import user.*;
 
-import static io.restassured.RestAssured.given;
-
 public class CreateUserTest {
 
     CreateUserData createUser = new CreateUserData();
     User user = createUser.random();
 
+    User existingUser = createUser.existingUser();
+
     UserRequest request = new UserRequest();
     UserResponse response = new UserResponse();
+    private String accessToken;
 
-    public String accessToken;
-
-    // Создание нового пользователя
-    @Before
-    public void successfulCreationOfNewUser() {
-        ValidatableResponse createUser = request.successfulСreation(user);
+    @DisplayName("Проверить создание нового пользователя")
+    @Test
+    public void successfulCreateOfNewUser() {
+        ValidatableResponse createUser = request.successfulCreate(user);
         accessToken = response.successCreate(createUser);
     }
 
     @DisplayName("Проверить создание пользователя, который уже зарегистрирован")
     @Test
-    public void creatingUserWithExistingData() {
-        ValidatableResponse createUser = request.successfulСreation(user);
+    public void createUserWithExistingData() {
+        ValidatableResponse createUser = request.successfulCreate(existingUser);
         response.errorWhenCreating2(createUser);
     }
 
     @DisplayName("Проверить создание пользователя и не заполнить одно из обязательных полей")
     @Description("Проверяем запрос без обязательного поля Name")
     @Test
-    public void creatingNewUserWithoutNecessarilyField(){
+    public void createNewUserWithoutNecessarilyField(){
         IncompleteUser incompleteUser = new IncompleteUser(user.getEmail(), user.getPassword());
         ValidatableResponse failCreateUser = request.errorWhenCreating(incompleteUser);
-        response.errorWhenCreating(failCreateUser);
+        accessToken = response.errorWhenCreating(failCreateUser);
     }
 
     // Удаление созданого пользователя
     @After
     public void deleteUser(){
-        ValidatableResponse deleteUser = request.deleteUser(accessToken);
-        response.successDeleteUser(deleteUser);
+        if(accessToken != null) {
+            request.deleteUser(accessToken);
+        }
     }
 }
